# contracts/openapi/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: 宝宝成长相机 REST API
  version: 1.0.0
  description: |
    端 ↔ 后端 REST 契约，对齐 [design-api.md](../../docs/design-api.md)。
    大版本 URI 前缀 `v1`；非破坏性变更仅增字段，破坏性变更升至 `v2`。
  contact:
    name: baobao API
  license:
    name: Proprietary
servers:
  - url: https://api-cn.babygrowth.app
    description: 中国区域
  - url: https://api-os.babygrowth.app
    description: 海外区域
security:
  - bearerAuth: []
...
paths:
...
  /v1/ai/plays:
    get:
      tags:
        - ai
      summary: 玩法目录
      operationId: aiListPlays
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 玩法列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /v1/ai/tasks/{taskId}/appeal:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - ai
      summary: 申诉（仅 rejected）
      operationId: aiAppealTask
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 申诉已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /v1/feeds/family:
    get:
      tags:
        - feed
      summary: 家庭 Feed
      operationId: feedListFamily
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Feed 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /v1/credits/balance:
    get:
      tags:
        - credits
      summary: 当前余额
      operationId: creditsGetBalance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 余额
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /v1/credits/transactions:
    get:
      tags:
        - credits
      summary: 账本流水（分页）
      operationId: creditsListTransactions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 流水列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /v1/credits/rates:
    get:
      tags:
        - credits
      summary: 各玩法当前积分单价
      operationId: creditsGetRates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 单价列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /v1/subscriptions/products:
    get:
      tags:
        - subscriptions
      summary: SKU 列表
      operationId: subscriptionsListProducts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: SKU 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required:
        - code
        - requestId
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: ok
        requestId:
          type: string
          example: req_01HZX
        data:
          type: object
          additionalProperties: true
...
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...